+++ frontends/web/src/routes/account/send/send.tsx
@@ -237,6 +237,7 @@ class Send extends Component<Props, State> {
       customFee: this.state.customFee,
       sendAll: (this.state.sendAll ? 'yes' : 'no'),
       selectedUTXOs: Object.keys(this.selectedUTXOs),
+      paymentRequest: null,
     };
   };